+++ go/core/pkg/sandboxbackend/channel_helpers/channel_helpers_test.go
@@ -1,4 +1,4 @@
-package channels
+package channel_helpers
 
 import (
 	"context"
